+++ Makefile
@@ -3,7 +3,7 @@
 venv:
 	python3.8 -m venv venv
 	./venv/bin/pip install --upgrade pip
-	./venv/bin/pip install -e art-common/ -e doozer/ -e elliott/ -e pyartcd/
+	./venv/bin/pip install -e doozer/ -e elliott/ -e pyartcd/
 	./venv/bin/pip install -r doozer/requirements-dev.txt -r pyartcd/requirements-dev.txt
 	cd elliott && ../venv/bin/pip install '.[tests]'
 	# source venv/bin/activate
