--- art-common/README.md
@@ -1,1 +1,0 @@
-# commonlib - ART common library
