+++ artcommon/constants.py
