+++ doozer/artcommon
@@ -1,0 +1,1 @@
+../artcommon
