+++ doozer/doozerlib/cli/release_gen_assembly.py
@@ -14,7 +14,7 @@ from doozerlib.cli import cli, pass_runtime, click_coroutine
 from doozerlib import exectools
 from doozerlib.model import Model
 from doozerlib import brew
-from artcommonlib import rhcos
+from artcommon import rhcos
 from doozerlib.rpmcfg import RPMMetadata
 from doozerlib.image import BrewBuildImageInspector
 from doozerlib.runtime import Runtime
