+++ doozer/doozerlib/cli/scan_sources.py
@@ -6,7 +6,7 @@ import yaml
 from typing import List, Tuple, Optional
 from github import GithubException
 
-from artcommonlib.rhcos import get_primary_container_name
+from artcommon.rhcos import get_primary_container_name
 from doozerlib import brew, exectools, rhcos, util
 from doozerlib.cli import cli, pass_runtime
 from doozerlib.cli import release_gen_payload as rgp
