+++ elliott/elliottlib/model.py
@@ -1,1 +1,1 @@
-from artcommonlib.model import ModelException, MissingModel, Missing, ListModel, Model
+from artcommon.model import ModelException, MissingModel, Missing, ListModel, Model
