+++ install.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-pip3 install -e art-common/ -e doozer/ -e elliott/ -e pyartcd/
+pip3 install -e doozer/ -e elliott/ -e pyartcd/
