+++ pyartcd/pyartcd/pipelines/operator_sdk_sync.py
@@ -5,7 +5,7 @@ import requests
 import click
 import koji
 from errata_tool import Erratum
-from artcommonlib.constants import BREW_DOWNLOAD_URL, BREW_HUB
+from artcommon.constants import BREW_DOWNLOAD_URL, BREW_HUB
 from pyartcd import constants, util, jenkins
 from pyartcd.cli import cli, click_coroutine, pass_runtime
 from pyartcd.runtime import Runtime
